+++ app.js
@@ -1,4 +1,5 @@
 const express = require('express')
+const constants = require('./server/util/constants')
 const app = express()
 
 //Settings
@@ -8,6 +9,7 @@ app.set('port', 8080)
 app.use(express.json())
 
 //Routes
-app.use(require('./server/microservices/recipe/routes'))
+app.use(constants.API_PATH, require('./server/microservices/recipe/routes'))
+app.use(constants.API_PATH, require('./server/microservices/ingredient/routes'))
 
 module.exports = app