+++ server/microservices/ingredient/routes.js
@@ -3,5 +3,9 @@ const controller = require('./controller')
 
 const router = express.Router();
 
+router.get('/ingredient', controller.getAllIngredients)
+router.post('/ingredient', controller.saveIngredient)
+router.get('/ingredient/:ID', controller.getIngredientById)
+router.put('/ingredient', controller.updateIngredient)
 
 module.exports = router