+++ CHANGELOG.md
@@ -8,6 +8,10 @@ This file is a contains a list of changes for each release.
 * Add `PUT method` to update a selected recipe
 * Add `GET by ID method` to get a recipe by ID
 * Create `mapper.js` to convert a JSON to an object
+* Add `GET method` to get all the ingredients stored in the database
+* Add `POST method` to save an ingredient to the database
+* Add `PUT method` to update a selected ingredient
+* Add `GET by ID method` to get an ingredient by its ID
 
 ## v0.2.0 (20-12-2021)